issue #207: relocate device capability interfaces to outbound ports, guard with ArchUnit
