+++ core/src/main/kotlin/org/agrfesta/sh/api/core/application/ports/outbounds/devices/Inspectable.kt
@@ -1,4 +1,4 @@
-package org.agrfesta.sh.api.core.domain.devices
+package org.agrfesta.sh.api.core.application.ports.outbounds.devices
 
 import arrow.core.Either
 import org.agrfesta.sh.api.core.domain.failures.DevicesProviderFailure
