+++ core/src/main/kotlin/org/agrfesta/sh/api/core/application/ports/outbounds/devices/ProviderDevicesFactory.kt
@@ -1,7 +1,6 @@
 package org.agrfesta.sh.api.core.application.ports.outbounds.devices
 
 import org.agrfesta.sh.api.core.domain.devices.Device
-import org.agrfesta.sh.api.core.domain.devices.DeviceDriver
 import org.agrfesta.sh.api.core.domain.devices.Provider
 
 interface ProviderDevicesFactory {
