+++ core/src/main/kotlin/org/agrfesta/sh/api/core/application/ports/outbounds/devices/Sensor.kt
@@ -1,6 +1,7 @@
-package org.agrfesta.sh.api.core.domain.devices
+package org.agrfesta.sh.api.core.application.ports.outbounds.devices
 
 import arrow.core.Either
+import org.agrfesta.sh.api.core.domain.devices.SensorReadings
 
 interface Sensor : DeviceDriver {
     fun fetchReadings(): Either<SensorReadingsFailure, SensorReadings>
