+++ core/src/main/kotlin/org/agrfesta/sh/api/core/application/ports/outbounds/devices/SensorReadingsFailure.kt
@@ -1,0 +1,5 @@
+package org.agrfesta.sh.api.core.application.ports.outbounds.devices
+
+interface SensorReadingsFailure
+
+data class FailureByException(val reason: Exception) : SensorReadingsFailure
