+++ core/src/main/kotlin/org/agrfesta/sh/api/core/application/ports/outbounds/devices/SharedHeater.kt
@@ -1,0 +1,5 @@
+package org.agrfesta.sh.api.core.application.ports.outbounds.devices
+
+interface Heater : OnOffActuator
+
+interface SharedHeater : Heater
