+++ core/src/main/kotlin/org/agrfesta/sh/api/core/application/usecases/FetchSensorReadingsService.kt
@@ -5,16 +5,16 @@ import arrow.core.getOrElse
 import arrow.core.left
 import arrow.core.right
 import org.agrfesta.sh.api.core.application.ports.inbounds.FetchSensorReadingsUseCase
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.BatteryPowered
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DeviceBatteryRepository
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DeviceDriver
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DevicesRepository
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.ProviderDevicesFactory
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.Sensor
 import org.agrfesta.sh.api.core.application.ports.outbounds.home.HomeStateRefreshPublisher
 import org.agrfesta.sh.api.core.application.ports.outbounds.sensors.SensorsCurrentReadingsRepository
-import org.agrfesta.sh.api.core.domain.devices.BatteryPowered
 import org.agrfesta.sh.api.core.domain.devices.Device
-import org.agrfesta.sh.api.core.domain.devices.DeviceDriver
 import org.agrfesta.sh.api.core.domain.devices.DeviceFeature
-import org.agrfesta.sh.api.core.domain.devices.Sensor
 import org.agrfesta.sh.api.core.domain.devices.ThermoHygroDataValue
 import org.agrfesta.sh.api.core.domain.failures.FetchSensorReadingsError
 import org.agrfesta.sh.api.core.domain.failures.FetchSensorReadingsFailure
