+++ core/src/main/kotlin/org/agrfesta/sh/api/core/domain/devices/ActuatorStatus.kt
@@ -1,0 +1,3 @@
+package org.agrfesta.sh.api.core.domain.devices
+
+enum class ActuatorStatus { ON, OFF, UNDEFINED }
