--- core/src/main/kotlin/org/agrfesta/sh/api/core/domain/devices/DeviceDriver.kt
@@ -1,7 +1,0 @@
-package org.agrfesta.sh.api.core.domain.devices
-
-import java.util.*
-
-interface DeviceDriver : DeviceProviderIdentity {
-    val uuid: UUID
-}
