--- core/src/main/kotlin/org/agrfesta/sh/api/core/domain/devices/SharedHeater.kt
@@ -1,5 +1,0 @@
-package org.agrfesta.sh.api.core.domain.devices
-
-interface Heater : OnOffActuator
-
-interface SharedHeater : Heater
