+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/netatmo/devices/NetatmoDevicesFactory.kt
@@ -1,9 +1,9 @@
 package org.agrfesta.sh.api.providers.netatmo.devices
 
 import org.agrfesta.sh.api.core.application.ports.outbounds.TimeProvider
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DeviceDriver
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.ProviderDevicesFactory
 import org.agrfesta.sh.api.core.domain.devices.Device
-import org.agrfesta.sh.api.core.domain.devices.DeviceDriver
 import org.agrfesta.sh.api.core.domain.devices.Provider
 import org.agrfesta.sh.api.providers.netatmo.ConditionalOnNetatmo
 import org.agrfesta.sh.api.providers.netatmo.NetatmoClient
