+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/switchbot/devices/SwitchBotDevicesFactory.kt
@@ -1,8 +1,8 @@
 package org.agrfesta.sh.api.providers.switchbot.devices
 
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DeviceDriver
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.ProviderDevicesFactory
 import org.agrfesta.sh.api.core.domain.devices.Device
-import org.agrfesta.sh.api.core.domain.devices.DeviceDriver
 import org.agrfesta.sh.api.core.domain.devices.Provider
 import org.agrfesta.sh.api.providers.switchbot.ConditionalOnSwitchBot
 import org.agrfesta.sh.api.providers.switchbot.SwitchBotDevicesClient
