+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/switchbot/devices/SwitchBotMeter.kt
@@ -5,14 +5,14 @@ import arrow.core.left
 import arrow.core.right
 import com.fasterxml.jackson.databind.JsonNode
 import kotlinx.coroutines.runBlocking
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.BatteryPowered
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.FailureByException
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.Inspectable
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.Sensor
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.SensorReadingsFailure
 import org.agrfesta.sh.api.core.domain.commons.Temperature
-import org.agrfesta.sh.api.core.domain.devices.BatteryPowered
-import org.agrfesta.sh.api.core.domain.devices.FailureByException
-import org.agrfesta.sh.api.core.domain.devices.Inspectable
 import org.agrfesta.sh.api.core.domain.devices.Provider
-import org.agrfesta.sh.api.core.domain.devices.Sensor
 import org.agrfesta.sh.api.core.domain.devices.SensorReadings
-import org.agrfesta.sh.api.core.domain.devices.SensorReadingsFailure
 import org.agrfesta.sh.api.core.domain.failures.DevicesProviderError
 import org.agrfesta.sh.api.core.domain.failures.DevicesProviderFailure
 import org.agrfesta.sh.api.providers.switchbot.SwitchBotDevicesClient
