+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/switchbot/devices/SwitchBotMiniHub.kt
@@ -1,6 +1,6 @@
 package org.agrfesta.sh.api.providers.switchbot.devices
 
-import org.agrfesta.sh.api.core.domain.devices.DeviceDriver
+import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DeviceDriver
 import org.agrfesta.sh.api.core.domain.devices.Provider
 import java.util.*
